feat(matrix): add util for getting users
